requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
google-api-python-client==1.6.5
google-auth==1.3.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
oauthlib==2.0.6
pkg-resources==0.0.0
praw==6.0.0
prawcore==1.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
update-checker==0.16
uritemplate==3.0.0
urllib3==1.22
